Compute annual degree day accumulation at listed x,y coordinate locations

A function that computes the total annual degree day accumulation for Lycorma delicatula development, based on monthly mean temperature data. This function was developed from the work of Dennis Calvin on the development of populations of Lycorma delicatula living on Ailanthus altissima at different DD accumulations.

compute_DD_suitability(x, thresh = "adult_emergence_1")

Arguments

x

Data import. Should be a .csv file or data frame that contains input temperature data. If a .csv file, file path should be in the format produced by the file.path() function (i.e. with '/' instead of '\'). See details for additional import formatting.

thresh

A numeric value or pre-build character string, which represents a commonly used threshold (see details for list of pre-build values). This sets the threshold for the number of Degree Days necessary for a population of Lycorma delicatula to reach the desired life stage.

Value

A data frame containing the total annual degree days and a binary measure of that location's suitability for establishment by Lycorma delicatula, per coordinate pair. It will have the same length as the input and will contain the following columns:

line no. = line number

cell = grid cell unique identifier

x = longitude

y = latitude

total_annual_dds = total degree days in one calendar year for that coordinate pair

suitability = binary, 1 if total_annual_dds >= thresh, 0 if total_annual_dds < thresh

Error Messages

An error message will be given if the input data x are in the incorrect format (see details) or if an incompatible value is given for parameter thresh (if value is not a preset character string or a numeric object).

Details

x:

Data frames / .csv files should contain 15 columns. The initial 3 columns should contain descriptors, with the 1st column as the raster cell number (see examples) and the 2nd and 3rd columns as longitude and latitude, respectively. If 1st column is unavailable, see "troubleshooting section of examples for code to produce this. Columns 4-15 should contain temperature data for each month of the year (one month per column).

thresh:

We have pre-built a number of important DD thresholds, as outlined by Dennis Calvin's work. These pre-build values can be called by entering their name as a character string. Pre-built values for thresh include the following:

"adult_emergence_1" = 991 DD (default value), when 1% of a population reaches the adult life stage. This was outlined as a critical threshold for establishment in North America.

"adult_emergence_50" = 1115.5 DD, when 50% of a population reaches the adult stage.

"egg_laying_10" = 1564.1 DD, when 10% of a population reaches egg-laying.

"egg_laying_50" = 1616.4 DD, when 50% of a population reaches egg-laying

Examples